+++ integration-tests/set-code-hash/lib.rs
@@ -73,7 +73,7 @@ pub mod incrementer {
         type E2EResult<T> = std::result::Result<T, Box<dyn std::error::Error>>;
 
         #[ink_e2e::test(additional_contracts = "./updated-incrementer/Cargo.toml")]
-        async fn set_code_works(mut client: ink_e2e::Client<C, E>) -> E2EResult<()> {
+        async fn set_code_works<Client: E2EBackend>(mut client: Client) -> E2EResult<()> {
             // Given
             let constructor = IncrementerRef::new();
             let contract = client