+++ integration-tests/wildcard-selector/lib.rs
@@ -69,8 +69,8 @@ pub mod wildcard_selector {
         }
 
         #[ink_e2e::test]
-        async fn arbitrary_selectors_handled_by_wildcard(
-            mut client: ink_e2e::Client<C, E>,
+        async fn arbitrary_selectors_handled_by_wildcard<Client: E2EBackend>(
+            mut client: Client,
         ) -> E2EResult<()> {
             // given
             let constructor = WildcardSelectorRef::new();
@@ -122,8 +122,8 @@ pub mod wildcard_selector {
         }
 
         #[ink_e2e::test]
-        async fn wildcard_complement_works(
-            mut client: ink_e2e::Client<C, E>,
+        async fn wildcard_complement_works<Client: E2EBackend>(
+            mut client: Client,
         ) -> E2EResult<()> {
             // given
             let constructor = WildcardSelectorRef::new();